requirements.txt
aerich==0.3.3
aiomysql==0.0.20
aiosqlite==0.16.0
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
ecdsa==0.14.1
fastapi==0.61.2
h11==0.11.0
iso8601==0.1.13
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.2
PyMySQL==0.9.2
PyPika==0.44.0
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.17
typing-extensions==3.7.4.3
uvicorn==0.12.2
